requirements.txt
# 多智能体通信与协作实验依赖包

# 核心依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# 多智能体环境
gym>=0.26.0
pygame>=2.5.0

# SMAC环境 (StarCraft II)
# git+https://github.com/oxwhirl/smac.git

# MPE环境 (Multi-Agent Particle Environment)
# git+https://github.com/openai/multiagent-particle-envs.git

# 工具包
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
jupyter>=1.0.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0

# 可选：GPU支持
# torch-cuda  # 根据CUDA版本选择

# 可选：强化学习工具
# stable-baselines3>=2.0.0
# ray[rllib]>=2.5.0